Enabled error logging, set up vote restrictions
Each user can now only vote one time per post. However, the visual
aspects of it need to be altered accordingly; it will appear as though
the user can vote as much as they want until they refresh the page.

This restriction will later be improved upon to allow the user to change
votes, but the key functionality is now in place.