Users can now submit!
Simplified submission.php, created a separate file for the submission
form, added category dropdown.

The submissions are successully being uploaded to our database if they
pass all requirements specified in submit.php, namely user must be
logged in, post must be less than or equal to 300 characters, and a
category must be selected.