Submissions are now working for the moment
The verification testing code has been removed for now, though. This
leaves a minor possibility for people to make a script to easily vote on
a million submissions, but I believe I know how I will fix this issue:

Don't use a verification code. Instead, every time the user loads a
page, add the ID's of the loaded submissions to the user's session and
unload any that may now be invalid. Through this, the vote.php script
can verify whether it's possible for the user to actually be able to
vote on the particular submission.